feat(ios): support for Fabric  - move specs outside the ts build scope to avoid compiling the Turbo / Native components code
