+++ specs/NativeComponentCircle.ts
