+++ specs/NativeComponentGooglePolygon.ts
