+++ specs/NativeComponentMapView.ts
