+++ specs/NativeComponentOverlay.ts
