+++ specs/NativeComponentPolyline.ts
