+++ src/createFabricMap.tsx
@@ -7,8 +7,8 @@ import type {
   EdgePadding,
   SnapshotOptions,
 } from './MapView.types';
-import NativeAirMapsModule, {MapBoundaries} from './specs/NativeAirMapsModule';
-import {MapFabricNativeProps} from './specs/NativeComponentMapView';
+import NativeAirMapsModule, {MapBoundaries} from '../specs/NativeAirMapsModule';
+import {MapFabricNativeProps} from '../specs/NativeComponentMapView';
 
 export type FabricMapViewProps = MapFabricNativeProps;
